Allow dumping serialized struct fields
It's possible to have a data type that can represent the same value
in more than one way. For such cases sometimes it's useful to have
the same hash.

This commit adds new "method" tag in "hash" group. Its value is
used to call a specific method of a struct field and dump its return
value instead of the field itself. If method is not found, or returns
more than one value, it panics.